Fix warning: narrowing conversion from 'size_t' (aka 'unsigned long') to signed type 'int32_t' (aka 'int') is implementation-defined [cppcoreguidelines-narrowing-conversions]

+++ hash-programs/cch.hpp
@@ -294,7 +294,7 @@ public:
     }
 
     /// Get the size (in bytes) of the state.
-    [[nodiscard]] constexpr static int get_state_size_bytes() noexcept { return sizeof(state_); }
+    [[nodiscard]] constexpr static int get_state_size_bytes() noexcept { return static_cast<int>(sizeof(state_)); }
 
     /// Get the maximum number of digest bytes that can be returned.
     [[nodiscard]] constexpr static int get_max_digest_size_bytes() noexcept { return get_state_size_bytes() / 4; }
